requirements.txt
streamlit==1.28.0
pandas
gspread
google-auth
